Moved the example plugin to seperate file
